+++ examples/sharded_array_write_read.rs
@@ -29,7 +29,7 @@ fn sharded_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
         chrono::Utc::now().format("[%T%.3f] ").to_string()
     });
     let store_readable_listable = usage_log.create_readable_listable_transformer(store.clone());
-    let store = usage_log.create_writable_transformer(store);
+    let store = usage_log.create_readable_writable_transformer(store);
 
     // Create a group and write metadata to filesystem
     let group_path = "/group";