+++ src/array/array_builder.rs
@@ -309,6 +309,9 @@ impl ArrayBuilder {
             additional_fields: self.additional_fields.clone(),
             parallel_codecs: self.parallel_codecs,
             include_zarrs_metadata: true,
+            chunk_locks: parking_lot::Mutex::default(),
+            #[cfg(feature = "async")]
+            async_chunk_locks: async_lock::Mutex::default(),
         })
     }
 }