+++ src/storage/storage_transformer/usage_log.rs
@@ -11,16 +11,17 @@ use crate::{
     metadata::Metadata,
     storage::{
         ListableStorage, ListableStorageTraits, ReadableListableStorage, ReadableStorage,
-        ReadableStorageTraits, StorageError, StoreKey, StoreKeyRange, StoreKeyStartValue,
-        StoreKeys, StoreKeysPrefixes, StorePrefix, WritableStorage, WritableStorageTraits,
+        ReadableStorageTraits, ReadableWritableStorage, ReadableWritableStorageTraits,
+        StorageError, StoreKey, StoreKeyRange, StoreKeyStartValue, StoreKeys, StoreKeysPrefixes,
+        StorePrefix, WritableStorage, WritableStorageTraits,
     },
 };
 
 #[cfg(feature = "async")]
 use crate::storage::{
     AsyncListableStorage, AsyncListableStorageTraits, AsyncReadableListableStorage,
-    AsyncReadableStorage, AsyncReadableStorageTraits, AsyncWritableStorage,
-    AsyncWritableStorageTraits,
+    AsyncReadableStorage, AsyncReadableStorageTraits, AsyncReadableWritableStorageTraits,
+    AsyncWritableStorage, AsyncWritableStorageTraits,
 };
 
 use super::StorageTransformerExtension;
@@ -64,6 +65,13 @@ impl StorageTransformerExtension for UsageLogStorageTransformer {
     }
 
     fn create_readable_transformer<'a>(&self, storage: ReadableStorage<'a>) -> ReadableStorage<'a> {
+        self.create_transformer(storage)
+    }
+
+    fn create_readable_writable_transformer<'a>(
+        &'a self,
+        storage: ReadableWritableStorage<'a>,
+    ) -> ReadableWritableStorage<'a> {
         self.create_transformer(storage)
     }
 
@@ -281,6 +289,11 @@ impl<TStorage: ?Sized + WritableStorageTraits> WritableStorageTraits
     }
 }
 
+impl<TStorage: ?Sized + ReadableWritableStorageTraits> ReadableWritableStorageTraits
+    for UsageLogStorageTransformerImpl<TStorage>
+{
+}
+
 #[cfg(feature = "async")]
 #[cfg_attr(feature = "async", async_trait::async_trait)]
 impl<TStorage: ?Sized + AsyncReadableStorageTraits> AsyncReadableStorageTraits
@@ -441,4 +454,11 @@ impl<TStorage: ?Sized + AsyncWritableStorageTraits> AsyncWritableStorageTraits
         )?;
         self.storage.erase_prefix(prefix).await
     }
+}
+
+#[cfg(feature = "async")]
+#[cfg_attr(feature = "async", async_trait::async_trait)]
+impl<TStorage: ?Sized + AsyncReadableWritableStorageTraits> AsyncReadableWritableStorageTraits
+    for UsageLogStorageTransformerImpl<TStorage>
+{
 }