+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - Added `ReadableWritableStorage` 
+ - Added `{async_}store_set_partial_values`
+ - **Breaking** Added `create_readable_writable_transformer` to `StorageTransformerExtension` trait
+
+### Changed
+ - **Breaking**  `ReadableStorageTraits` is no longer a supertrait of `WritableStorageTraits`
+   - `WritableStorage` is no longer implicitly readable. Use `ReadableWritableStorage`
+ - **Breaking**: `{Async}WritableStorageTraits::set_partial_values()` no longer include default implementations
+   - Use new `{async_}store_set_partial_values` utility functions instead
+ - Add `#[must_use]` to `Array::builder`, `Array::chunk_grid_shape`, and `ArrayBuilder::from_array`
+
 ## [0.7.3] - 2023-12-22
 
 ### Added